+++ src/main/java/com/sprint/mission/findex/domain/indexdata/dto/IndexDataResponse.java
@@ -1,7 +1,6 @@
 package com.sprint.mission.findex.domain.indexdata.dto;
 
-import com.sprint.mission.findex.domain.indexdata.entity.IndexData;
-import com.sprint.mission.findex.domain.indexinfo.entity.IndexInfo.SourceType;
+import com.sprint.mission.findex.domain.indexinfo.entity.SourceType;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
